[Gepardec/mega#657] removed not-zep-functions from ZepService
